@@ -3,40 +3,33 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "dev": "nuxt",
-    "build": "nuxt build",
-    "start": "nuxt start",
-    "generate": "nuxt generate",
+    "dev": "nuxi dev",
+    "build": "nuxi build",
+    "start": "nuxi start",
+    "generate": "nuxi generate",
     "lint:js": "eslint --ext .js,.vue --ignore-path .gitignore .",
     "lint": "npm run lint:js"
   },
   "dependencies": {
-    "@nuxt/typescript-runtime": "^2.0.0",
-    "@nuxtjs/axios": "^5.12.5",
-    "@nuxtjs/composition-api": "^0.26.0",
-    "@nuxtjs/i18n": "^7.0.2",
-    "@nuxtjs/pwa": "^3.3.3",
-    "@nuxtjs/sitemap": "^2.4.0",
-    "chrome-aws-lambda": "^5.5.0",
     "core-js": "^3.6.5",
-    "nuxt": "^2.15.7",
-    "puppeteer-core": "^5.5.0"
+    "nuxt": "^3.10.0"
   },
   "devDependencies": {
-    "@nuxt/types": "^2.14.6",
-    "@nuxt/typescript-build": "^2.0.3",
-    "@nuxtjs/eslint-config": "^3.1.0",
-    "@nuxtjs/eslint-config-typescript": "^3.0.0",
-    "@nuxtjs/eslint-module": "^2.0.0",
-    "@nuxtjs/tailwindcss": "^3.1.0",
+    "@nuxt/types": "2.17.3",
+    "@nuxtjs/eslint-config": "^12.0.0",
+    "@nuxtjs/eslint-config-typescript": "^12.1.0",
+    "@nuxtjs/eslint-module": "^4.1.0",
+    "@nuxtjs/i18n": "^8.0.1",
+    "@nuxtjs/sitemap": "^5.1.0",
+    "@nuxtjs/tailwindcss": "^6.11.2",
+    "@vite-pwa/nuxt": "^0.5.0",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.10.0",
-    "eslint-config-prettier": "^6.12.0",
-    "eslint-plugin-nuxt": "^1.0.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "postcss": "7.0.34",
-    "postcss-nested": "4.2.3",
-    "prettier": "^2.1.2",
-    "puppeteer": "^5.5.0"
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-nuxt": "^4.0.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "postcss": "8.4.35",
+    "postcss-nested": "6.0.1",
+    "prettier": "^3.2.5"
   }
 }
