@@ -1,24 +1,19 @@
 <template>
   <div class="form__display-checkbox checkbox">
     <label class="checkbox__label">
-      <input
-        class="checkbox__input"
-        type="checkbox"
-        :checked="displaySection"
-        @change="
-          changeDisplaySection({ sectionName, status: $event.target.checked })
-        "
-      />
+      <input class="checkbox__input" type="checkbox" :checked="displaySection" @change="
+        changeDisplaySection({ sectionName, status: ($event.target as HTMLInputElement).checked })
+        " />
       <span class="checkbox__text"> {{ $t('display-section') }} </span>
     </label>
   </div>
 </template>
 
 <script lang="ts">
-import Vue from 'vue';
+import { defineComponent } from 'vue';
 import { useCvState } from '~/data/useCvState';
 
-export default Vue.extend({
+export default defineComponent({
   props: {
     sectionName: {
       type: String,
