@@ -1,44 +1,26 @@
 <template>
   <div class="flex w-full gap-3">
-    <input
-      id="languages"
-      v-model.trim="tagInputLang.lang"
-      class="form__control mt-2 mb-1"
-      type="text"
-      placeholder="Spanish"
-      @keyup.enter="updateSkill"
-    />
-    <select
-      id="level"
-      v-model="tagInputLang.level"
-      name="level"
-      class="form__control mt-2 mb-1"
-      type="selet"
-      @keyup.enter="updateSkill"
-    >
+    <input id="languages" v-model.trim="tagInputLang.lang" class="form__control mt-2 mb-1" type="text"
+      placeholder="Spanish" @keyup.enter="updateSkill" />
+    <select id="level" v-model="tagInputLang.level" name="level" class="form__control mt-2 mb-1" type="selet"
+      @keyup.enter="updateSkill">
       <option v-for="level in LEVELS" :key="level" :value="level">
         {{ $t(level) }}
       </option>
     </select>
-    <button
-      class="form__btn"
-      type="button"
-      :disabled="tagInputLangEmpty"
-      :aria-disabled="tagInputLangEmpty"
-      aria-live="assertive"
-      @click="updateSkill"
-    >
+    <button class="form__btn" type="button" :disabled="tagInputLangEmpty" :aria-disabled="tagInputLangEmpty"
+      aria-live="assertive" @click="updateSkill">
       {{ $t('add') }}
     </button>
   </div>
 </template>
 <script lang="ts">
-import { reactive, computed, toRefs } from '@nuxtjs/composition-api';
-import Vue from 'vue';
+import { reactive, computed, toRefs } from 'vue';
+import { defineComponent } from 'vue';
 import { useCvState } from '~/data/useCvState';
-import { Level, LEVELS } from '~/types/cvfy';
+import { type Level, LEVELS } from '~/types/cvfy';
 
-export default Vue.extend({
+export default defineComponent({
   name: 'CvInputLang',
   props: {
     tagListName: {
